# documentation/api-docs/reference/openapi.yaml
openapi: 3.1.0
info:
  title: OpenAPI definition
  version: "1.0.0"
  description: |
      API for profession orientation and quiz-based assessment.

      ## Versioning Strategy
      This API follows URI-based semantic versioning:
      - Major versions are encoded in the base path: /api/v1.0
      - Backward-incompatible changes trigger a new major version (v2.0)
      - Backward-compatible changes are released within the same version
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
- url: http://localhost:8082
  description: Generated server url
security:
- BearerAuth: []
paths:
...
  /questions/{id}/order/{ord}:
    put:
      tags:
      - question-controller
      summary: Update question order
      description: Changes the display order of a question within a quiz version.
      operationId: updateOrder
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
          format: int32
      - name: ord
        in: path
        required: true
        schema:
          type: integer
          format: int32
      responses:
        "401":
          description: Unauthorized
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ExceptionDto"
        "400":
          description: Bad Request
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ExceptionDto"
        "404":
          description: Question not found
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ExceptionDto"
        "500":
          description: Internal Server Error
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ExceptionDto"
        "200":
          description: Order updated
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/QuestionDto"
        "403":
          description: Forbidden
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/QuestionDto"
...
  /options/{optionId}/traits:
    put:
      tags:
      - option-trait-controller
      summary: Update traits for an option
      description: |
        Updates the list of traits associated with a specific option.
        Traits not present in the request will be removed.
        Accessible only to administrators.
      operationId: updateTraits
      parameters:
      - name: optionId
        in: path
        required: true
        schema:
          type: integer
          format: int32
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/OptionTraitListRequest"
        required: true
      responses:
        "401":
          description: Unauthorized
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ExceptionDto"
        "400":
          description: Invalid request payload
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ExceptionDto"
        "404":
          description: Option not found
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ExceptionDto"
        "500":
          description: Internal Server Error
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ExceptionDto"
        "200":
          description: Traits successfully updated
        "403":
          description: Forbidden
    post:
      tags:
      - option-trait-controller
      summary: Assign traits to an option
      description: |
        Assigns a list of traits with weights to a specific option.
        Existing trait assignments (if any) will be replaced.
        Accessible only to administrators.
      operationId: assignTraits
      parameters:
      - name: optionId
        in: path
        required: true
        schema:
          type: integer
          format: int32
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/OptionTraitListRequest"
        required: true
      responses:
        "401":
          description: Unauthorized
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ExceptionDto"
        "400":
          description: Invalid request payload
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ExceptionDto"
        "404":
          description: Option not found
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ExceptionDto"
        "500":
          description: Internal Server Error
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ExceptionDto"
        "200":
          description: Traits successfully assigned
        "403":
          description: Forbidden
  /options/{id}:
    put:
      tags:
      - option-controller
      summary: Update an option
      description: |
        Updates an existing answer option.
        Only administrators can modify options.
      operationId: update_5
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
          format: int32
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateOptionRequest"
        required: true
      responses:
        "401":
          description: Unauthorized
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ExceptionDto"
        "400":
          description: Validation error
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ExceptionDto"
        "404":
          description: Option not found
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ExceptionDto"
        "500":
          description: Internal Server Error
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ExceptionDto"
        "200":
          description: Option updated
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/OptionDto"
        "403":
          description: Forbidden
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/OptionDto"
    delete:
      tags:
      - option-controller
      summary: Delete an option
      description: |
        Deletes an answer option permanently.
        This operation cannot be undone.
      operationId: delete_5
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
          format: int32
      responses:
        "401":
          description: Unauthorized
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ExceptionDto"
        "400":
          description: Bad Request
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ExceptionDto"
        "404":
          description: Option not found
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ExceptionDto"
        "500":
          description: Internal Server Error
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ExceptionDto"
        "200":
          description: Option deleted
        "403":
          description: Forbidden
  /options/{id}/order/{ord}:
    put:
      tags:
      - option-controller
      summary: Update option order
      description: |
        Updates the display order of an option within a question.
        Order must be a non-negative integer.
      operationId: updateOrder_1
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
          format: int32
      - name: ord
        in: path
        required: true
        schema:
          type: integer
          format: int32
      responses:
        "401":
          description: Unauthorized
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ExceptionDto"
        "400":
          description: Invalid order value
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ExceptionDto"
        "404":
          description: Option not found
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ExceptionDto"
        "500":
          description: Internal Server Error
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ExceptionDto"
        "200":
          description: Option order updated
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/OptionDto"
        "403":
          description: Forbidden
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/OptionDto"
...
  /options:
    post:
      tags:
      - option-controller
      summary: Create an answer option
      description: |
        Creates a new answer option for a question.
        Only administrators are allowed to perform this operation.
      operationId: create_5
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateOptionRequest"
        required: true
      responses:
        "401":
          description: Unauthorized
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ExceptionDto"
        "400":
          description: Validation error
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ExceptionDto"
        "404":
          description: Not Found
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ExceptionDto"
        "500":
          description: Internal Server Error
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ExceptionDto"
        "200":
          description: Option successfully created
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/OptionDto"
        "403":
          description: Forbidden
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/OptionDto"
...
components:
  schemas:
    ExceptionDto:
      type: object
      description: Standardized API error response returned for all handled exceptions
      properties:
        code:
          type: integer
          format: int32
          description: HTTP status code representing the error type
          examples:
          - 400
        time:
          type: string
          format: date-time
          description: Timestamp indicating when the error occurred (ISO-8601 format)
          examples:
          - 2025-01-15T14:32:10Z
        message:
          type: object
          description: |
            Error details. May be:
            - a string message (e.g. 'Invalid credentials')
            - a structured object containing validation errors
          examples:
          - Invalid email or password
          oneOf:
          - type: string
          - {}
...
    OptionDto:
      type: object
      description: Represents a selectable option for a quiz question
      properties:
        id:
          type: integer
          format: int32
          description: Unique identifier of the option
          examples:
          - 54
        questionId:
          type: integer
          format: int32
          description: Identifier of the question this option belongs to
          examples:
          - 7
        ord:
          type: integer
          format: int32
          description: Order index of the option within the question (1-based)
          examples:
          - 1
          minimum: 1
        label:
          type: string
          description: Text label shown to the user for this option
          examples:
          - Strongly agree
    QuestionDto:
      type: object
      description: Represents a question within a quiz version
      properties:
        id:
          type: integer
          format: int32
          description: Unique identifier of the question
          examples:
          - 101
        quizVersionId:
          type: integer
          format: int32
          description: Identifier of the quiz version this question belongs to
          examples:
          - 1
        ord:
          type: integer
          format: int32
          description: Order of the question within the quiz
          examples:
          - 3
          minimum: 1
        qtype:
          type: string
          description: Type of question defining how the user should answer
          enum:
          - single_choice
          - multi_choice
          - liker_scale_5
          - liker_scale_7
          examples:
          - single_choice
        text:
          type: string
          description: Text displayed to the user for this question
          examples:
          - I enjoy solving practical problems using my hands
        options:
          type: array
          description: List of selectable answer options for the question
          items:
            $ref: "#/components/schemas/OptionDto"
...
    OptionTraitListRequest:
      type: object
      description: Request payload for assigning multiple trait weights to a question
        option
      properties:
        traits:
          type: array
          description: List of trait-weight associations for the option
          items:
            $ref: "#/components/schemas/OptionTraitRequest"
      required:
      - traits
    OptionTraitRequest:
      type: object
      description: Represents the association between an option and a trait with a
        weighted contribution
      properties:
        traitId:
          type: integer
          format: int32
          description: Identifier of the trait profile
          examples:
          - 3
        weight:
          type: number
          description: |
            Weight applied to the trait when this option is selected.
            Can be positive or negative depending on scoring logic.
          examples:
          - 1.25
      required:
      - traitId
      - weight
    UpdateOptionRequest:
      type: object
      description: Request payload for updating a quiz option
      properties:
        ord:
          type: integer
          format: int32
          description: New display order of the option
          examples:
          - 2
          minimum: 1
        label:
          type: string
          description: Updated label text of the option
          examples:
          - Strongly agree
...
    CreateOptionRequest:
      type: object
      description: Request payload for creating a new question option
      properties:
        questionId:
          type: integer
          format: int32
          description: Identifier of the question this option belongs to
          examples:
          - 101
        ord:
          type: integer
          format: int32
          description: Display order of the option within the question
          examples:
          - 1
          minimum: 1
        label:
          type: string
          description: Displayed label of the option
          examples:
          - Strongly agree
...
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT